More random stuff WIP Rylobot

+++ projects/Rylogic/gui/graph_control.cs
@@ -2318,7 +2318,7 @@ namespace pr.gui
 		/// <summary>Plot a moving average curve over the data</summary>
 		private void PlotMovingAverage(Graphics gfx, Series.RdrOptions opts, Matrix3x3 g2c, Iterator<GraphValue> iter)
 		{
-			var ema = new ExpMovingAvr((uint)opts.MAWindowSize);
+			var ema = new ExpMovingAvr(opts.MAWindowSize);
 			using (var ma_pen = new Pen(opts.MALineColour, opts.MALineWidth))
 			{
 				bool first = true;
